Move mock engine into testing.
