+++ modules/light-clients/08-wasm/types/types_test.go
@@ -41,7 +41,7 @@ type TypesTestSuite struct {
 	testifysuite.Suite
 	coordinator *ibctesting.Coordinator
 	chainA      *ibctesting.TestChain
-	mockVM      *types.MockWasmEngine
+	mockVM      *wasmtesting.MockWasmEngine
 
 	ctx      sdk.Context
 	store    storetypes.KVStore
@@ -95,7 +95,7 @@ func (suite *TypesTestSuite) SetupWasmWithMockVM() {
 }
 
 func (suite *TypesTestSuite) setupWasmWithMockVM() (ibctesting.TestingApp, map[string]json.RawMessage) {
-	suite.mockVM = types.NewMockWasmEngine()
+	suite.mockVM = wasmtesting.NewMockWasmEngine()
 	// TODO: move default functionality required for wasm client testing to the mock VM
 	suite.mockVM.StoreCodeFn = func(code wasmvm.WasmCode) (wasmvm.Checksum, error) {
 		hash := sha256.Sum256(code)
@@ -119,7 +119,7 @@ func (suite *TypesTestSuite) setupWasmWithMockVM() (ibctesting.TestingApp, map[s
 	suite.mockVM.RegisterQueryCallback(types.StatusMsg{}, func(codeID wasmvm.Checksum, env wasmvmtypes.Env, queryMsg []byte, store wasmvm.KVStore, goapi wasmvm.GoAPI, querier wasmvm.Querier, gasMeter wasmvm.GasMeter, gasLimit uint64, deserCost wasmvmtypes.UFraction) ([]byte, uint64, error) {
 		resp, err := json.Marshal(types.StatusResult{Status: exported.Active.String()})
 		suite.Require().NoError(err)
-		return resp, types.DefaultGasUsed, nil
+		return resp, wasmtesting.DefaultGasUsed, nil
 	})
 
 	db := dbm.NewMemDB()
